chore(Globalstyles): moves test utils helper to test-utils folder
